+++ src/main/java/fr/alasdiablo/jerintegration/JERIntegration.java
@@ -1,8 +1,6 @@
 package fr.alasdiablo.jerintegration;
 
 import fr.alasdiablo.jerintegration.compat.CompatibilityHandler;
-import net.minecraftforge.common.MinecraftForge;
-import net.minecraftforge.event.entity.player.PlayerEvent;
 import net.minecraftforge.fml.ModList;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
@@ -16,29 +14,21 @@ public class JERIntegration {
     public static final  Logger               LOGGER      = LogManager.getLogger(JERIntegration.MOD_ID);
     private static final ModList              MOD_LIST    = ModList.get();
     public final         CompatibilityHandler compatibilityHandler;
-    public               boolean              oneTimeLoad = false;
 
     public JERIntegration() {
         this.compatibilityHandler = new CompatibilityHandler();
         FMLJavaModLoadingContext.get().getModEventBus().addListener(this::setup);
-        MinecraftForge.EVENT_BUS.addListener(this::onPlayerJoinWorld);
     }
 
     private void setup(final FMLCommonSetupEvent commonSetupEvent) {
-
-    }
-
-    private void onPlayerJoinWorld(PlayerEvent.PlayerLoggedInEvent event) {
-        if (!this.oneTimeLoad) {
-            this.compatibilityHandler.preInit();
-            this.compatibilityHandler.init();
-            this.oneTimeLoad = true;
-        }
+        this.compatibilityHandler.init();
     }
 
     public static class Compat {
         public static boolean AE2                   = MOD_LIST.isLoaded("ae2");
         public static boolean CREATE                = MOD_LIST.isLoaded("create");
         public static boolean IMMERSIVE_ENGINEERING = MOD_LIST.isLoaded("immersiveengineering");
+        public static boolean TINKERS_CONSTRUCT     = MOD_LIST.isLoaded("tconstruct");
+        public static boolean MEKANISM              = MOD_LIST.isLoaded("mekanism");
     }
 }